LayoutTests/accessibility/mac/ruby-hierarchy-roles-expected.txt

This tests that the ruby containers are exposed with the appropriate hierarchy and roles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rubyElem != null is true
PASS axRuby != null is true
PASS role is expectedRubyRole
PASS subrole is expectedRubyInlineSubrole
PASS axRubyBase != null is true
PASS role is expectedRubyRole
PASS subrole is expectedRubyBaseSubrole
PASS axRubyText != null is true
PASS role is expectedRubyRole
PASS subrole is expectedRubyTextSubrole
PASS axRuby != null is true
PASS role is expectedRubyRole
PASS subrole is expectedRubyBlockSubrole
PASS role is expectedRubyRole
PASS subrole is expectedRubyInlineSubrole
PASS axRubyBase != null is true
PASS role is expectedRubyRole
PASS subrole is expectedRubyBaseSubrole
PASS axRubyText != null is true
PASS role is expectedRubyRole
PASS subrole is expectedRubyTextSubrole
PASS successfullyParsed is true

TEST COMPLETE
basetext rubytext
